+++ pyoffers/exceptions.py
@@ -3,3 +3,7 @@
 
 class HasOffersException(BaseException):
     pass
+
+
+class MaxRetriesExceeded(HasOffersException):
+    pass